requirements.txt
opencv-contrib-python
pysimplegui
pyserial
scikit-learn
matplotlib
pandas
numpy
requests
tqdm
openvino